Add security checks on standards controller

+++ Arkitektum.Orden/Controllers/Api/StandardsController.cs
@@ -4,6 +4,7 @@ using Arkitektum.Orden.Data;
 using Arkitektum.Orden.Models;
 using Arkitektum.Orden.Models.Api;
 using Arkitektum.Orden.Services;
+using Arkitektum.Orden.Utils;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -14,11 +15,13 @@ namespace Arkitektum.Orden.Controllers.Api
     [Route("api/standards")]
     public class StandardsController : BaseController
     {
+        private readonly IApplicationService _applicationService;
         private readonly IStandardService _standardService;
 
-        public StandardsController(ISecurityService securityService, IStandardService standardService) : base(
+        public StandardsController(ISecurityService securityService, IApplicationService applicationService, IStandardService standardService) : base(
             securityService)
         {
+            _applicationService = applicationService;
             _standardService = standardService;
         }
 
@@ -32,6 +35,10 @@ namespace Arkitektum.Orden.Controllers.Api
         [HttpPost("application")]
         public async Task<IActionResult> AddStandardToApplication([FromBody] ApplicationStandardViewModel model)
         {
+            Application application = await _applicationService.GetAsync(model.ApplicationId);
+            if (!_securityService.CurrrentUserHasAccessToApplication(application, AccessLevel.Write))
+                return Forbid();
+            
             await _standardService.AddStandardToApplication(new ApplicationStandard().Map(model));
             
             return NoContent();
@@ -40,6 +47,10 @@ namespace Arkitektum.Orden.Controllers.Api
         [HttpDelete("{standardId}/application/{applicationId}")]
         public async Task<IActionResult> RemoveStandardFromApplication(int standardId, int applicationId)
         {
+            Application application = await _applicationService.GetAsync(applicationId);
+            if (!_securityService.CurrrentUserHasAccessToApplication(application, AccessLevel.Write))
+                return Forbid();
+            
             await _standardService.RemoveStandardFromApplication(
                 new ApplicationStandard() {StandardId = standardId, ApplicationId = applicationId});
             return NoContent();
